Clear progress bar text when hiding it
When files are uploaded the progress bar text is set accordingly.
However, other operations that show the progress bar, like deleting
files, do not explicitly set any text. Due to that, when the progress
bar was shown again after uploading files the text did not match the
operation. To solve that now the text is cleared when the progress bar
is hidden (it is not cleared when it is shown as in some cases the text
is set already before showing the progress bar).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>